src/pyproject.toml
[project]
name = "meshwiki"
version = "0.1.0"
description = "A modern wiki platform inspired by MoinMoin, Graphingwiki, and Obsidian"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2>=3.1",
    "markdown>=3.5",
    "pymdown-extensions>=10.0",
    "pydantic-settings>=2.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0",
    "itsdangerous>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "pytest-playwright>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["meshwiki"]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["meshwiki"]

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]
